fix: fix nft collection page loading when switch tab
