+++ src/app/(normal)/nft/[address]/page.tsx
@@ -1,17 +1,14 @@
 'use client';
 
-import { t } from '@lingui/macro';
 import { TextOverflowTooltip } from '@masknet/theme';
+import { ChainId } from '@masknet/web3-shared-evm';
 import { useQuery } from '@tanstack/react-query';
 import { notFound } from 'next/navigation.js';
-import { useState } from 'react';
 
 import ComeBack from '@/assets/comeback.svg';
 import { CollectionInfo } from '@/components/CollectionDetail/CollectionInfo.js';
-import { NFTList } from '@/components/CollectionDetail/NFTList.js';
-import { TopCollectors } from '@/components/CollectionDetail/TopCollectors.js';
+import { CollectionTabs } from '@/components/CollectionDetail/CollectionTabs.js';
 import { Loading } from '@/components/Loading.js';
-import { Tab, Tabs } from '@/components/Tabs/index.js';
 import type { SourceInURL } from '@/constants/enum.js';
 import { useComeBack } from '@/hooks/useComeback.js';
 import { useNFTFloorPrice } from '@/hooks/useNFTFloorPrice.js';
@@ -24,19 +21,10 @@ export default function Page({
     params: { address: string };
     searchParams: { source: SourceInURL; chainId: string };
 }) {
-    const tabs = [
-        {
-            label: t`Items`,
-            value: 'items',
-        },
-        {
-            label: t`Top Collectors`,
-            value: 'topCollectors',
-        },
-    ] as const;
-    const chainId = searchParams.chainId ? Number.parseInt(searchParams.chainId as string, 10) : undefined;
+    const chainId: ChainId | undefined = searchParams.chainId
+        ? Number.parseInt(searchParams.chainId as string, 10)
+        : undefined;
     const comeback = useComeBack();
-    const [currentTab, setCurrentTab] = useState<(typeof tabs)[number]['value']>('items');
     const { address } = params;
     const { data, isLoading, error } = useQuery({
         queryKey: ['collection-info', address, chainId],
@@ -75,23 +63,7 @@ export default function Page({
                     chainId={chainId}
                 />
             ) : null}
-            <div className="px-3 pb-3">
-                <Tabs value={currentTab} onChange={setCurrentTab} variant="second">
-                    {tabs.map((tab) => (
-                        <Tab value={tab.value} key={tab.value}>
-                            {tab.label}
-                        </Tab>
-                    ))}
-                </Tabs>
-                {
-                    {
-                        items: <NFTList address={address} chainId={chainId} />,
-                        topCollectors: (
-                            <TopCollectors address={address} totalQuantity={data?.total_quantity} chainId={chainId} />
-                        ),
-                    }[currentTab]
-                }
-            </div>
+            <CollectionTabs address={address} chainId={chainId} totalQuantity={data?.total_quantity} />
         </div>
     );
 }
